[Feature] support flush to SLS with zstd compress type (#527)
* [feature] support flush to SLS with zstd compress type

+++ core/config_manager/ConfigManager.cpp
@@ -187,8 +187,8 @@ Json::Value& ConfigManager::CheckPluginProcessor(Json::Value& pluginConfigJson, 
         && (pluginConfigJson["processors"].isObject() || pluginConfigJson["processors"].isArray())) {
         // patch enable_log_position_meta to split processor if exists ...
         if (rootConfigJson["advanced"] && rootConfigJson["advanced"]["enable_log_position_meta"]) {
-            for (int i = 0; i < pluginConfigJson["processors"].size(); i++) {
-                Json::Value& processorConfigJson = pluginConfigJson["processors"][i];
+            for (size_t i = 0; i < pluginConfigJson["processors"].size(); i++) {
+                Json::Value& processorConfigJson = pluginConfigJson["processors"][int(i)];
                 if (processorConfigJson["type"] == "processor_split_log_string"
                     || processorConfigJson["type"] == "processor_split_log_regex") {
                     if (processorConfigJson["detail"]) {
@@ -197,7 +197,7 @@ Json::Value& ConfigManager::CheckPluginProcessor(Json::Value& pluginConfigJson, 
                     }
                     break;
                 }
-                pluginConfigJson["processors"][i] = processorConfigJson;
+                pluginConfigJson["processors"][int(i)] = processorConfigJson;
             }
         }
     }
